Add files via upload
This project centers on analysis the statistics of the players as well as of Arsenal football club as a whole to gain actionable insights. Through meticulous data visualisation,creation of new dataset,importing and exporting data and analysis,it aims to determine key performance metrics ,factors affecting performances of player and comparison of output of crucial players across corresponding seasons.It makes use of dbms to make modifications in data , find similarities between different datasets and export the same.